[REFACTOR] Extracts the creation logic of the employers body to a factory
